% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_pcp2.R
\name{gather_pcp2}
\alias{gather_pcp2}
\title{gather for parallel coordinate plot}
\usage{
gather_pcp2(data, ...)
}
\arguments{
\item{data}{The data set used}

\item{...}{choose the columns to be used}
}
\value{
dataframe with gathered data and some extra columns
}
\description{
It should be used before using other drawing functions
}
