% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pcp_band.R
\name{geom_pcp_band}
\alias{geom_pcp_band}
\title{Bands for the parallel coordinate plot}
\usage{
geom_pcp_band(
  mapping = NULL,
  data = NULL,
  stat = "pcpband",
  position = "identity",
  ...,
  freespace = 0.1,
  boxwidth = 0,
  rugwidth = 0,
  interwidth = 1,
  breakpoint = NULL,
  reverse = FALSE,
  merge = FALSE,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by [aes()] or
[aes_()]. If specified and `inherit.aes = TRUE` (the
default), it is combined with the default mapping at the top level of the
plot. You must supply `mapping` if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If `NULL`, the default, the data is inherited from the plot
   data as specified in the call to [ggplot()].

   A `data.frame`, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   [fortify()] for which variables will be created.

   A `function` will be called with a single argument,
   the plot data. The return value must be a `data.frame`, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`colour = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.
#'}

\item{freespace}{The total gap space among levels within each factor variable.}

\item{boxwidth}{The width of the boxes for factor variables, either a scalar or a vector with length equal to the number of factor variables.}

\item{rugwidth}{The width of the rugs for numeric variables,either a scalar or a vector with length equal to the number of numeric variables.}

\item{interwidth}{The width for the lines between every neighboring variables, either
a scalar or a vector with length equal to the total number of variables subtracting one.}

\item{breakpoint}{Positions indicating where to break, can be a vector.
To break three or more factor variables to better show the relations between adjacent factor variables.
Can't be used when there is only one or two factor variables.}

\item{reverse}{reverse the plot, useful especially when you want to reverse the structure in factor blocks,
i.e. to become more ordered from right to left}

\item{merge}{To merge the bands or not}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
`NA`, the default, includes if any aesthetics are mapped.
`FALSE` never includes, and `TRUE` always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If `FALSE`, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. [borders()].}
}
\description{
To add bands from factor to factor
}
\examples{
library(ggplot2)
library(dplyr)
data(mtcars)
mtcars \%>\%
mutate(cyl = factor(cyl),
      vs = factor(vs),
      am = factor(am),
      gear = factor(gear)) \%>\%
 ggplot(aes(vars = vars(cyl, vs:gear))) +
 geom_pcp(aes(color = vs), boxwidth = 0.2, breakpoint = 2:3) +
 geom_pcp_box(boxwidth = 0.2) +
 geom_pcp_band(boxwidth = 0.2, breakpoint = 2:3) +
 geom_pcp_text(boxwidth = 0.2)
}
