% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-pcp-axes.r, R/geom-pcp-boxes.r,
%   R/geom-pcp.r, R/geom_pcp_labels.r
\docType{data}
\name{GeomPcp_axes}
\alias{GeomPcp_axes}
\alias{GeomPcp_boxes}
\alias{GeomPcp}
\alias{GeomPcp_labels}
\title{Proto version of the pcp geoms}
\description{
These functions are only exported so that they are visible to
the ggplot2 internal functions.
User-relevant documentation can be found instead in \code{\link[=geom_pcp]{geom_pcp()}}.
}
\keyword{datasets}
