% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpcp-package.R
\docType{package}
\name{ggpcp-package}
\alias{ggpcp}
\alias{ggpcp-package}
\title{ggpcp: Parallel Coordinate Plots in the 'ggplot2' Framework}
\description{
Modern Parallel Coordinate Plots have been introduced in the 1980s as a way to visualize arbitrarily many numeric variables. This Grammar of Graphics implementation also incorporates categorical variables into the plots in a principled manner. By separating the data managing part from the visual rendering, we give full access to the users while keeping the number of parameters manageably low.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heike/ggpcp}
  \item Report bugs at \url{https://github.com/heike/ggpcp/issues}
}

}
\author{
\strong{Maintainer}: Heike Hofmann \email{hofmann@iastate.edu} (\href{https://orcid.org/0000-0001-6216-5183}{ORCID})

Authors:
\itemize{
  \item Susan VanderPlas \email{susan.vanderplas@unl.edu} (\href{https://orcid.org/0000-0002-3803-0972}{ORCID})
  \item Yawei Ge \email{yaweige@iastate.edu}
}

}
\keyword{internal}
