% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatPcp.R
\name{stat_pcp}
\alias{stat_pcp}
\title{Parallel coordinate plot for both numeric and categorical data}
\usage{
stat_pcp(mapping = NULL, data = NULL, geom = "segment",
  position = "identity", ..., freespace = 0.1, boxwidth = 0,
  rugwidth = 0, interwidth = 1, breaks = NULL, breakwidth = 0.1,
  na.rm = FALSE, show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by [aes()] or
[aes_()]. If specified and `inherit.aes = TRUE` (the
default), it is combined with the default mapping at the top level of the
plot. You must supply `mapping` if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If `NULL`, the default, the data is inherited from the plot
   data as specified in the call to [ggplot()].

   A `data.frame`, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   [fortify()] for which variables will be created.

   A `function` will be called with a single argument,
   the plot data. The return value must be a `data.frame`, and
   will be used as the layer data.}

\item{geom}{The geometric object to use display the data}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`colour = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}

\item{freespace}{The total gap space among levels within each factor variable}

\item{boxwidth}{The width of the box for each factor variable}

\item{rugwidth}{The width of the rugs for numeric variable}

\item{interwidth}{The width for the lines between every neighboring variables, either
a scalar or a vector.}

\item{breaks}{To break three or more factors into peices}

\item{breakwidth}{will be removed}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
`NA`, the default, includes if any aesthetics are mapped.
`FALSE` never includes, and `TRUE` always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If `FALSE`, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. [borders()].}
}
\description{
The parallel coordinate plot displays multiple y-axes, and shows the observations across
several dimensions as ploi-lines. This function work well with both numeric and categorical
variables at the same time after proper scaling.
}
