% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_pcp.R
\name{transform_pcp}
\alias{transform_pcp}
\title{Transforming values in a parallel coordinate plot}
\usage{
transform_pcp(data, method = "minmax")
}
\arguments{
\item{data}{data frame as returned by gather_pcp}

\item{method}{string specifying the method that should be used for scaling the values
in a parallel coordinate plot. One of "minmax' (default),}
}
\description{
Transforming values in a parallel coordinate plot
}
