% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pcp.R
\name{geom_pcp}
\alias{geom_pcp}
\title{Parallel coordinate plot for both numeric and categorical data}
\usage{
geom_pcp(mapping = NULL, data = NULL, stat = "pcp",
  position = "identity", ..., method = "uniminmax", freespace = 0.1,
  boxwidth = 0, rugwidth = 0, interwidth = 1, breakpoint = NULL,
  overplot = "original", arrow = NULL, arrow.fill = NULL,
  lineend = "butt", linejoin = "round", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by [aes()] or
[aes_()]. If specified and `inherit.aes = TRUE` (the
default), it is combined with the default mapping at the top level of the
plot. You must supply `mapping` if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If `NULL`, the default, the data is inherited from the plot
   data as specified in the call to [ggplot()].

   A `data.frame`, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   [fortify()] for which variables will be created.

   A `function` will be called with a single argument,
   the plot data. The return value must be a `data.frame`, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{...}{Other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`colour = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.
#'}

\item{method}{which method should be used to transform the values of each variable into acommon y axis? See `transform_pcp` for details.}

\item{freespace}{The total gap space among levels within each factor variable.}

\item{boxwidth}{The width of the boxes for factor variables, either a scalar or a vector with length equal to the number of factor variables.}

\item{rugwidth}{The width of the rugs for numeric variables,either a scalar or a vector with length equal to the number of numeric variables.}

\item{interwidth}{The width for the lines between every neighboring variables, either
a scalar or a vector with length equal to the total number of variables subtracting one.}

\item{breakpoint}{Positions indicating where to break, can be a vector.
To break three or more factor variables to better show the relations between adjacent factor variables.
Can't be used when there is only one or two factor vairbles.}

\item{overplot}{methods used to conduct overplotting when overplotting becomes an issue.}

\item{arrow}{specification for arrow heads, as created by arrow()}

\item{arrow.fill}{fill colour to use for the arrow head (if closed). NULL means use colour aesthetic}

\item{lineend}{Line end style (round, butt, square)}

\item{linejoin}{Line join style (round, mitre, bevel)}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
`NA`, the default, includes if any aesthetics are mapped.
`FALSE` never includes, and `TRUE` always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If `FALSE`, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. [borders()].}
}
\description{
The parallel coordinate plot displays multiple y-axes, and shows the observations across
several dimensions as lines. This function work well with both numeric and categorical
variables at the same time after proper scaling.
}
