% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom2plotly.r
\name{plotly_helpers}
\alias{plotly_helpers}
\alias{to_basic.GeomPcp}
\alias{to_basic.GeomPcp_axes}
\alias{to_basic.GeomPcp_labels}
\alias{to_basic.GeomPcp_boxes}
\alias{prefix_class}
\title{Plotly helpers}
\usage{
\method{to_basic}{GeomPcp}(data, prestats_data, layout, params, p, ...)

\method{to_basic}{GeomPcp_axes}(data, prestats_data, layout, params, p, ...)

\method{to_basic}{GeomPcp_labels}(data, prestats_data, layout, params, p, ...)

\method{to_basic}{GeomPcp_boxes}(data, prestats_data, layout, params, p, ...)

prefix_class(x, y)
}
\description{
Helper functions to make it easier to automatically create plotly charts
}
\examples{
# example code
flea_plotly <- highlight_key(flea_pcp, ~pcp_id)
p <- flea_plotly  |>
  ggplot(aes_pcp()) +
  geom_pcp_axes() +
  geom_pcp(aes(colour = species, label = pcp_id))
pp <- ggplotly(p)
highlight(pp, on="plotly_hover")

}
\keyword{internal}
