% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_pcp.R
\name{transform_pcp}
\alias{transform_pcp}
\title{Transforming values in a parallel coordinate plot}
\usage{
transform_pcp(data, method = "uniminmax")
}
\arguments{
\item{data}{data frame as returned by gather_pcp}

\item{method}{string specifying the method that should be used for scaling the values
in a parallel coordinate plot (see Details).}
}
\description{
Set of transformations to use in parallel coordinate plots. All transformation aim for y values in the interval [0,1].
}
\details{
\code{method} is a character string that denotes how to scale the variables
in the parallel coordinate plot. Options are named in the same way as the options in `ggparcoord` (GGally):
\itemize{
  \item{\code{raw}}{: raw data used, no scaling will be done.}
  \item{\code{std}}{: univariately, subtract mean and divide by standard deviation. To get values into a [0,1] interval we use a linear transformation of f(y) = y/4+0.5. }
  \item{\code{robust}}{: univariately, subtract median and divide by median absolute deviation. To get values into a [0,1] interval we use a linear transformation of f(y) = y/4+0.5. }
  \item{\code{uniminmax}}{: univariately, scale so the minimum of the variable is zero, and the maximum is one.}
  \item{\code{globalminmax}}{: gobal scaling; the global maximum is mapped to 1,
    global minimum across the variables is mapped to 0. }
}
}
